refactor: add font management module and update global types
