+++ src/index.html
@@ -8502,7 +8502,6 @@
     <script defer src="libs/polylabel.min.js?v1.105.0"></script>
     <script defer src="libs/lineclip.min.js?v1.105.0"></script>
     <script defer src="libs/simplify.js?v1.105.6"></script>
-    <script defer src="modules/fonts.js?v=1.99.03"></script>
     <script defer src="modules/ui/layers.js?v=1.111.0"></script>
     <script defer src="modules/ui/measurers.js?v=1.99.00"></script>
     <script defer src="modules/ui/style-presets.js?v=1.100.00"></script>
