+++ src/modules/index.ts
@@ -14,3 +14,4 @@ import "./zones-generator";
 import "./religions-generator";
 import "./provinces-generator";
 import "./ice";
+import "./fonts";
